@@ -1,78 +1,88 @@
 {
-  "name": "tailwind-nextjs-starter-blog",
-  "version": "2.4.0",
+  "name": "nextjs-notion-starter-kit",
+  "version": "2.0.0",
   "private": true,
+  "description": "The perfect starter kit for building beautiful websites with Next.js and Notion.",
+  "author": "Travis Fischer <travis@transitivebullsh.it>",
+  "repository": "transitive-bullshit/nextjs-notion-starter-kit",
+  "license": "MIT",
+  "packageManager": "pnpm@10.11.1",
+  "engines": {
+    "node": ">=18"
+  },
+  "type": "module",
   "scripts": {
-    "start": "next dev",
-    "dev": "cross-env INIT_CWD=$PWD next dev",
-    "build": "cross-env INIT_CWD=$PWD next build && cross-env NODE_OPTIONS='--experimental-json-modules' node ./scripts/postbuild.mjs",
-    "serve": "next start",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "deploy": "vercel deploy",
+    "deps:upgrade": "[ -z $GITHUB_ACTIONS ] && pnpm up -L notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:update on CI'",
+    "deps:link": "[ -z $GITHUB_ACTIONS ] && run-s deps:link:* || echo 'Skipping deps:update on CI'",
+    "deps:unlink": "[ -z $GITHUB_ACTIONS ] && pnpm add notion-client notion-types notion-utils react-notion-x || echo 'Skipping deps:update on CI'",
+    "deps:link:notion-types": "pnpm link ../react-notion-x/packages/notion-types",
+    "deps:link:notion-utils": "pnpm link ../react-notion-x/packages/notion-utils",
+    "deps:link:notion-client": "pnpm link ../react-notion-x/packages/notion-client",
+    "deps:link:react-notion-x": "pnpm link ../react-notion-x/packages/react-notion-x",
     "analyze": "cross-env ANALYZE=true next build",
-    "lint": "next lint --fix --dir pages --dir app --dir components --dir lib --dir layouts --dir scripts",
-    "prepare": "husky"
+    "analyze:server": "cross-env BUNDLE_ANALYZE=server next build",
+    "analyze:browser": "cross-env BUNDLE_ANALYZE=browser next build",
+    "test": "run-p test:*",
+    "test:lint": "eslint .",
+    "test:prettier": "prettier '**/*.{js,jsx,ts,tsx}' --check"
   },
   "dependencies": {
-    "@headlessui/react": "2.2.0",
-    "@next/bundle-analyzer": "15.2.4",
-    "@tailwindcss/forms": "^0.5.9",
-    "@tailwindcss/postcss": "^4.0.5",
-    "@tailwindcss/typography": "^0.5.15",
-    "body-scroll-lock": "^4.0.0-beta.0",
-    "contentlayer2": "0.5.5",
-    "esbuild": "0.25.2",
-    "github-slugger": "^2.0.0",
-    "gray-matter": "^4.0.2",
-    "hast-util-from-html-isomorphic": "^2.0.0",
-    "image-size": "2.0.1",
-    "next": "15.2.4",
-    "next-contentlayer2": "0.5.5",
-    "next-themes": "^0.4.6",
-    "pliny": "0.4.1",
-    "postcss": "^8.4.24",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "reading-time": "1.5.0",
-    "rehype-autolink-headings": "^7.1.0",
-    "rehype-citation": "^2.3.0",
-    "rehype-katex": "^7.0.0",
-    "rehype-katex-notranslate": "^1.1.4",
-    "rehype-preset-minify": "7.0.0",
-    "rehype-prism-plus": "^2.0.0",
-    "rehype-slug": "^6.0.0",
-    "remark": "^15.0.0",
-    "remark-gfm": "^4.0.0",
-    "remark-github-blockquote-alert": "^1.2.1",
-    "remark-math": "^6.0.0",
-    "tailwindcss": "^4.0.5",
-    "unist-util-visit": "^5.0.0"
+    "@fisch0920/use-dark-mode": "^2.4.0",
+    "@keyvhq/core": "^2.1.7",
+    "@keyvhq/redis": "^2.1.8",
+    "@react-icons/all-files": "^4.1.0",
+    "classnames": "^2.5.1",
+    "expiry-map": "^2.0.0",
+    "fathom-client": "^3.7.2",
+    "ky": "^1.8.1",
+    "lqip-modern": "^2.2.1",
+    "next": "^15.3.4",
+    "notion-client": "^7.4.2",
+    "notion-types": "^7.4.2",
+    "notion-utils": "^7.4.2",
+    "p-map": "^7.0.3",
+    "p-memoize": "^7.1.1",
+    "posthog-js": "^1.255.1",
+    "prismjs": "^1.30.0",
+    "react": "^19.1.0",
+    "react-body-classname": "^1.3.1",
+    "react-dom": "^19.1.0",
+    "react-notion-x": "^7.4.2",
+    "react-tweet": "^3.2.2",
+    "react-use": "^17.6.0",
+    "rss": "^1.2.2"
   },
   "devDependencies": {
-    "@eslint/eslintrc": "^3.2.0",
-    "@eslint/js": "^9.16.0",
-    "@svgr/webpack": "^8.0.1",
-    "@types/mdx": "^2.0.12",
-    "@types/react": "^19.0.8",
-    "@typescript-eslint/eslint-plugin": "^8.12.0",
-    "@typescript-eslint/parser": "^8.12.0",
+    "@fisch0920/config": "^1.1.4",
+    "@next/bundle-analyzer": "^15.3.4",
+    "@types/node": "^24.0.4",
+    "@types/prismjs": "^1.26.5",
+    "@types/react": "^19.1.8",
+    "@types/react-body-classname": "^1.1.10",
+    "@types/rss": "^0.0.32",
     "cross-env": "^7.0.3",
-    "eslint": "^9.14.0",
-    "eslint-config-next": "15.2.4",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.2.0",
-    "globals": "^15.12.0",
-    "husky": "^9.0.0",
-    "lint-staged": "^13.0.0",
-    "prettier": "^3.0.0",
-    "prettier-plugin-tailwindcss": "^0.6.11",
-    "typescript": "^5.1.3"
+    "eslint": "^9.29.0",
+    "npm-run-all2": "^8.0.4",
+    "prettier": "^3.6.1",
+    "typescript": "^5.8.3"
+  },
+  "overrides": {
+    "cacheable-request": {
+      "keyv": "npm:@keyvhq/core@~1.6.6"
+    }
+  },
+  "prettier": "@fisch0920/config/prettier",
+  "simple-git-hooks": {
+    "pre-commit": "npx lint-staged"
   },
   "lint-staged": {
-    "*.+(js|jsx|ts|tsx)": [
+    "*.{ts,tsx}": [
+      "prettier --ignore-unknown --write",
       "eslint --fix"
-    ],
-    "*.+(js|jsx|ts|tsx|json|css|md|mdx)": [
-      "prettier --write"
     ]
-  },
-  "packageManager": "yarn@3.6.1"
+  }
 }
